Move ProxyProvider to transport package
