+++ src/docs/asciidoc/tcp-client.adoc
@@ -454,7 +454,7 @@ The following example uses `ProxyProvider`:
 [source,java]
 ----
 import reactor.netty.Connection;
-import reactor.netty.tcp.ProxyProvider;
+import reactor.netty.transport.ProxyProvider;
 import reactor.netty.tcp.TcpClient;
 
 public class Application {
