+++ src/main/java/reactor/netty/http/client/HttpClientConnect.java
@@ -85,7 +85,7 @@ import reactor.netty.http.HttpOperations;
 import reactor.netty.http.HttpResources;
 import reactor.netty.resources.LoopResources;
 import reactor.netty.transport.AddressUtils;
-import reactor.netty.tcp.ProxyProvider;
+import reactor.netty.transport.ProxyProvider;
 import reactor.netty.tcp.SslProvider;
 import reactor.netty.tcp.TcpClient;
 import reactor.netty.channel.ChannelMetricsHandler;
