+++ src/main/java/reactor/netty/http/client/HttpClientDoOnError.java
@@ -37,7 +37,7 @@ import reactor.netty.ConnectionObserver;
 import reactor.netty.channel.BootstrapHandlers;
 import reactor.netty.http.Cookies;
 import reactor.netty.http.HttpOperations;
-import reactor.netty.tcp.ProxyProvider;
+import reactor.netty.transport.ProxyProvider;
 import reactor.netty.tcp.SslProvider;
 import reactor.netty.tcp.TcpClient;
 import reactor.util.context.Context;
