+++ src/main/java/reactor/netty/http/client/HttpClientOnConnectMap.java
@@ -23,7 +23,7 @@ import javax.annotation.Nullable;
 import io.netty.bootstrap.Bootstrap;
 import reactor.core.publisher.Mono;
 import reactor.netty.Connection;
-import reactor.netty.tcp.ProxyProvider;
+import reactor.netty.transport.ProxyProvider;
 import reactor.netty.tcp.SslProvider;
 import reactor.netty.tcp.TcpClient;
